requirements.txt
amqp==2.5.2
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
bleach==3.1.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
h5py==2.10.0
idna==2.9
importlib-metadata==0.23
ipykernel==5.1.3
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.11.1
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
kombu==4.6.8
lmdb==0.98
lxml==4.4.1
mahotas==1.4.9
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.18.2
opencv-contrib-python==4.2.0.32
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.4.2
pyparsing==2.4.4
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
redis==3.4.1
requests==2.23.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
terminado==0.8.2
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==0.6.0
gunicorn==19.9.0
